requirements.txt
aio-pika==9.0.5
aiomysql==0.2.0
aiormq==6.7.7
annotated-types==0.5.0
anyio==4.0.0
apache-skywalking==1.0.1
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
cryptography==3.4.8
Django==4.2.4
dnspython==2.4.2
ecdsa==0.14.1
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.0
fastapi-auth==0.1.1
greenlet==3.0.1
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
idna==3.4
loguru==0.7.2
motor==3.2.0
multidict==6.0.4
mysql-connector==2.2.9
packaging==23.2
pamqp==3.2.1
passlib==1.7.4
pika==1.3.2
prometheus-client==0.18.0
protobuf==4.25.1
psutil==5.9.7
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydantic_core==2.4.0
PyJWT==2.8.0
pymongo==4.4.1
PyMySQL==1.1.0
pyOpenSSL==21.0.0
python-jose==3.2.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.2
